# feat(op-acceptor): improved per-test logging