… op-acceptor/logging/filelogger_test.go
… op-acceptor/logging/json_parser_test.go
… op-acceptor/logging/templates/results.tmpl.html
… op-acceptor/runner/runner.go