**Description**

This change includes a number of improvements around test logging and ease of debugging failed tests:
* The 'tests' folder, and its confusing outputs are removed
* All test logs are placed in a file in either directory 'passed' or 'failed'
* All per-test logs are improved since now all lines are captured and included. Previously some were missing. (Example before/after below)
* A new, very very simple, HTML output summary is included which helps one to navigate to the pertinent log files (Pictured)

BEFORE:

```
--- FAIL: TestRPCConnectivity (5.27s)
    === RUN   TestRPCConnectivity
    rpc_connectivity_test.go:26: INFO [05-09|16:33:19.824] Started L2 RPC connectivity test
--- FAIL: TestRPCConnectivity (2.21s)
FAIL
FAIL    github.com/ethereum-optimism/optimism/op-acceptance-tests/tests/base    2.757s

FAIL    github.com/ethereum-optimism/optimism/op-acceptance-tests/tests/base    5.27s
```


AFTER:

```
=== RUN   TestRPCConnectivity
    rpc_connectivity_test.go:26: INFO [05-09|18:06:43.592] Started L2 RPC connectivity test
=== RUN   TestRPCConnectivity/L2_Chain_0
    rpc_connectivity_test.go:46: 
                Error Trace:    /Users/stefano/code/optimism/op-acceptance-tests/tests/base/rpc_connectivity_test.go:46
                                                        /Users/stefano/code/optimism/devnet-sdk/testing/systest/tb.go:51
                Error:          Not equal: 
                                expected: int(0)
                                actual  : *big.Int(2151908)
                Test:           TestRPCConnectivity/L2_Chain_0
                Messages:       L2 chain ID is not correct
--- FAIL: TestRPCConnectivity (2.18s)
    --- FAIL: TestRPCConnectivity/L2_Chain_0 (0.00s)
FAIL
FAIL    github.com/ethereum-optimism/optimism/op-acceptance-tests/tests/base    2.736s
```

<img width="1231" alt="Screenshot 2025-05-09 at 18 46 42" src="https://github.com/user-attachments/assets/d33d68d9-c569-4cc1-a632-081ce284c495" />



**Tests**

Ran against my local devnets.


**Metadata**

- Fixes #280  (please confirm, @sigma :) )